# tacocloud/tacocloud-api/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Taco Cloud Public API
  version: 1.0.0
  description: |
    Contrato oficial OpenAPI 3.0 para Taco Cloud (Laboratorio 6 - TC-35).
    Define los contratos de integración para ingredientes, diseño de tacos, pedidos con idempotencia
    y trazabilidad por Correlation ID, favoritos, cupones, pasarela de pago tokenizada, operaciones de cocina y
    anuncios operativos seguros. Incluye el modelo canónico de error RFC 7807 (ApiProblem).
  contact:
    name: Taco Cloud Engineering Team
    email: support@tacocloud.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de desarrollo local (v1 Oficial)
  - url: http://localhost:8080/api
    description: Servidor de desarrollo local (Alias legado - Deprecado)
...
paths:
...
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Consultar un pedido por su identificador
      operationId: getOrderById
      security:
        - basicAuth: []
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalle del pedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Pedido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
    delete:
      tags: [Orders]
      summary: Cancelar un pedido
      operationId: cancelOrder
      security:
        - basicAuth: []
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Pedido cancelado
...
  /users/me/orders:
    get:
      tags: [User Profile]
      summary: Consultar historial de pedidos del usuario autenticado
      operationId: getMyOrders
      security:
        - basicAuth: []
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Historial de pedidos paginado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderSummaryPageResponse'
  /users/me/orders/{id}:
    get:
      tags: [User Profile]
      summary: Consultar detalle de un pedido propio
      operationId: getMyOrderDetail
      security:
        - basicAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalle del pedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResponse'
        '404':
          description: Pedido no encontrado o no pertenece al usuario
  /users/me/favorites:
    get:
      tags: [User Profile]
      summary: Listar tacos marcados como favoritos por el usuario
      operationId: getMyFavorites
      security:
        - basicAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Lista de favoritos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FavoriteResponse'
  /users/me/favorites/{tacoId}:
    put:
      tags: [User Profile]
      summary: Agregar taco a favoritos
      operationId: addFavorite
      security:
        - basicAuth: []
        - bearerAuth: []
      parameters:
        - name: tacoId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Taco agregado a favoritos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteResponse'
    delete:
      tags: [User Profile]
      summary: Quitar taco de favoritos
      operationId: removeFavorite
      security:
        - basicAuth: []
        - bearerAuth: []
      parameters:
        - name: tacoId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Taco removido de favoritos
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT para usuarios autenticados
    basicAuth:
      type: http
      scheme: basic
      description: Autenticación básica HTTP (usuario y contraseña)
...
  schemas:
...
    IngredientResponse:
      type: object
      required:
        - id
        - name
        - type
      properties:
        id:
          type: string
          example: FLTO
        name:
          type: string
          example: Flour Tortilla
        type:
          type: string
          example: WRAP

    # --- TACOS ---
...
    TacoResponse:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          example: 64b8a1c9e4b0c23a8f123456
        name:
          type: string
          example: Carnitas Fiesta
        createdAt:
          type: string
          format: date-time
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientResponse'
        dietaryTags:
          type: array
          items:
            type: string
          example: [GLUTEN_FREE]
        allergens:
          type: array
          items:
            type: string
          example: [DAIRY]
        spiceLevel:
          type: string
          example: MEDIUM
...
    OrderItemResponse:
      type: object
      properties:
        taco:
          $ref: '#/components/schemas/TacoResponse'
        quantity:
          type: integer
          example: 2
        unitPrice:
          type: number
          format: double
          example: 4.50
        lineTotal:
          type: number
          format: double
          example: 9.00
...
    OrderResponse:
      type: object
      required:
        - id
        - deliveryName
        - total
      properties:
        id:
          type: string
          example: 64b8a1c9e4b0c23a8f999999
        placedAt:
          type: string
          format: date-time
        deliveryName:
          type: string
          example: Jane Doe
        deliveryStreet:
          type: string
          example: 123 Main St
        deliveryCity:
          type: string
          example: Austin
        deliveryState:
          type: string
          example: TX
        deliveryZip:
          type: string
          example: 78701
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        subtotal:
          type: number
          format: double
          example: 18.00
        discountAmount:
          type: number
          format: double
          example: 3.60
        total:
          type: number
          format: double
          example: 14.40
        currency:
          type: string
          example: USD
        status:
          type: string
          example: PLACED
        brand:
          type: string
          example: Visa
        last4:
          type: string
          example: '4242'
...
    OrderSummaryPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/OrderDetailResponse'
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          example: 5
    OrderDetailResponse:
      type: object
      properties:
        id:
          type: string
          example: 64b8a1c9e4b0c23a8f999999
        placedAt:
          type: string
          format: date-time
        deliveryName:
          type: string
          example: Jane Doe
        deliveryCity:
          type: string
          example: Austin
        total:
          type: number
          example: 14.40
        status:
          type: string
          example: READY

    # --- FAVORITOS ---
    FavoriteResponse:
      type: object
      properties:
        id:
          type: string
          example: fav-12345
        tacoId:
          type: string
          example: 64b8a1c9e4b0c23a8f123456
        taco:
          $ref: '#/components/schemas/TacoResponse'
        createdAt:
          type: string
          format: date-time

    # --- CUPONES ---
...
    ApiProblem:
      type: object
      required:
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          example: about:blank
        title:
          type: string
          example: Bad Request
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: Request validation failed
        instance:
          type: string
          example: /api/v1/orders
        code:
          type: string
          example: VALIDATION_ERROR
        violations:
          type: array
          items:
            $ref: '#/components/schemas/ApiProblemViolation'
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
    ApiProblemViolation:
      type: object
      properties:
        field:
          type: string
          example: deliveryName
        message:
          type: string
          example: Delivery name is required